+++ pkg/controllers/management/k3sbasedupgrade/template_test.go
@@ -1,4 +1,4 @@
-package k3sbaseupgrade
+package k3sbasedupgrade
 
 import "testing"
